python/requirements.txt
async-timeout==4.0.2
attrs==21.4.0
iniconfig==1.1.1
maturin==0.13.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.19.0
tomli==2.0.1
grpcio-tools==1.50.0
